test: move factories to core
